refactor out api functions into own file
