+++ app-web/plugins/gatsby-source-github-all/utils/sources/github/index.js
@@ -23,7 +23,7 @@ const { Base64 } = require('js-base64'); // eslint-disable-line
 const fetch = require('node-fetch'); // eslint-disable-line
 const ignore = require('ignore'); // eslint-disable-line
 const transformer = require('./githubFileTransformer');
-const { GITHUB_API_ENDPOINT, DEFUALT_IGNORES, GITHUB_SOURCE_SCHEMA } = require('../../constants');
+const { DEFUALT_IGNORES, GITHUB_SOURCE_SCHEMA } = require('../../constants');
 const {
   isConfigForFetchingAFile,
   isConfigForFetchingFiles,
@@ -33,57 +33,8 @@ const {
   filterFilesFromDirectories,
   applyBaseMetadata,
 } = require('./helpers');
+const { fetchFile, fetchGithubTree } = require('./api');
 const { validateSourcePropertiesAgainstSchema } = require('../../helpers');
-
-/**
- * Using the recursion param, this
- * function attempts to retrieve all directories/files from a repo
- * there is a limit on how deep it can go down a tree
- * as per https://developer.github.com/v3/git/trees/#get-a-tree-recursively
- * @param {String} repo
- * @param {String} owner
- * @param {String} token
- */
-const fetchGithubTree = async (repo, owner, token, branch = 'master') => {
-  const endPoint = `${GITHUB_API_ENDPOINT}/repos/${owner}/${repo}/git/trees/${branch}?recursive=1`;
-  try {
-    const result = await fetch(endPoint, {
-      method: 'GET',
-      headers: {
-        Authorization: `Bearer ${token}`,
-      },
-    });
-    const data = await result.json();
-    return data;
-  } catch (e) {
-    throw e;
-  }
-};
-/**
- * Fetches contents from file
- * note the media type header, it converts what would have been a
- * b64 encoded string of the file contents into either raw string data or json
- * @param {String} repo
- * @param {String} owner
- * @param {String} path
- * @param {String} token
- */
-const fetchFile = async (path, token) => {
-  try {
-    const result = await fetch(path, {
-      method: 'GET',
-      headers: {
-        Authorization: `Bearer ${token}`,
-        'X-GitHub-Media-Type': 'Accept: application/vnd.github.v3.raw+json',
-      },
-    });
-    const data = await result.json();
-    if (result.ok) return data;
-    return undefined;
-  } catch (e) {
-    throw e;
-  }
-};
 
 /**
  * returns array of fetch file promises
